+++ cmd/hla/hla_tcp/main.go
@@ -26,7 +26,7 @@ var (
 			WithDescription("set path to config, database files").
 			WithDefinedValue("."),
 		flag.NewFlagBuilder("-n", "--network").
-			WithDescription("set network key for connections").
+			WithDescription("set network key of connections from build").
 			WithDefinedValue(""),
 	).Build()
 )