+++ cmd/hls/main.go
@@ -26,11 +26,8 @@ var (
 			WithDescription("set path to config, database files").
 			WithDefinedValue("."),
 		flag.NewFlagBuilder("-n", "--network").
-			WithDescription("set network key for connections").
+			WithDescription("set network key of connections from build").
 			WithDefinedValue(""),
-		flag.NewFlagBuilder("-t", "--threads").
-			WithDescription("set num of parallel functions to calculate PoW").
-			WithDefinedValue("1"),
 	).Build()
 )