+++ internal/service/pkg/app/app_test.go
@@ -27,11 +27,8 @@ var (
 			WithDescription("set path to config, database files").
 			WithDefinedValue("."),
 		flag.NewFlagBuilder("-n", "--network").
-			WithDescription("set network key for connections").
+			WithDescription("set network key of connections from build").
 			WithDefinedValue(""),
-		flag.NewFlagBuilder("-t", "--threads").
-			WithDescription("set num of parallel functions to calculate PoW").
-			WithDefinedValue("1"),
 	).Build()
 )
 
@@ -70,11 +67,6 @@ func TestInitApp(t *testing.T) {
 		return
 	}
 
-	if _, err := InitApp([]string{"--path", tcTestdataPath, "--threads", "abc"}, tgFlags); err == nil {
-		t.Error("success init app with threads=abc")
-		return
-	}
-
 	if _, err := InitApp([]string{"--path", "./not_exist/path/to/hls"}, tgFlags); err == nil {
 		t.Error("success init app with undefined dir key")
 		return
@@ -110,7 +102,7 @@ func TestApp(t *testing.T) {
 	}
 
 	privKey := asymmetric.NewPrivKey()
-	app := NewApp(cfg, privKey, ".", 1)
+	app := NewApp(cfg, privKey, ".")
 
 	ctx, cancel := context.WithCancel(context.Background())
 	defer cancel()