+++ internal/service/pkg/app/config/config.go
@@ -22,6 +22,8 @@ type SConfigSettings struct {
 	FFetchTimeoutMS   uint64 `json:"fetch_timeout_ms" yaml:"fetch_timeout_ms"`
 	FQueuePeriodMS    uint64 `json:"queue_period_ms" yaml:"queue_period_ms"`
 	FWorkSizeBits     uint64 `json:"work_size_bits,omitempty" yaml:"work_size_bits,omitempty"`
+	FQBConsumers      uint64 `json:"qb_consumers,omitempty" yaml:"qb_consumers,omitempty"`
+	FPowParallel      uint64 `json:"pow_parallel,omitempty" yaml:"pow_parallel,omitempty"`
 	FNetworkKey       string `json:"network_key,omitempty" yaml:"network_key,omitempty"`
 }
 
@@ -106,6 +108,14 @@ func (p *SConfigSettings) GetQueuePeriod() time.Duration {
 
 func (p *SConfigSettings) GetNetworkKey() string {
 	return p.FNetworkKey
+}
+
+func (p *SConfigSettings) GetQBConsumers() uint64 {
+	return p.FQBConsumers
+}
+
+func (p *SConfigSettings) GetPowParallel() uint64 {
+	return p.FPowParallel
 }
 
 func (p *SConfig) GetSettings() IConfigSettings {