+++ internal/service/pkg/config/config.go
@@ -15,6 +15,8 @@ func GetConfigSettings(pCfg config.IConfig, pClient client.IClient) SConfigSetti
 			FNetworkKey:       sett.GetNetworkKey(),
 			FMessageSizeBytes: sett.GetMessageSizeBytes(),
 			FWorkSizeBits:     sett.GetWorkSizeBits(),
+			FQBConsumers:      sett.GetQBConsumers(),
+			FPowParallel:      sett.GetPowParallel(),
 			FFetchTimeoutMS:   uint64(sett.GetFetchTimeout() / time.Millisecond),
 			FQueuePeriodMS:    uint64(sett.GetQueuePeriod() / time.Millisecond),
 		},