+++ pkg/network/settings.go
@@ -23,6 +23,7 @@ type sSettings struct {
 type SSubSettings struct {
 	FLogger      gopeer_logger.ILogger
 	FParallel    uint64
+	FQBConsumers uint64
 	FServiceName string
 }
 
@@ -78,6 +79,14 @@ func (p *sSettings) useDefault() *sSettings {
 		p.FSubSettings.FServiceName = "_"
 	}
 
+	if p.FSubSettings.FParallel == 0 {
+		p.FSubSettings.FParallel = 1
+	}
+
+	if p.FSubSettings.FQBConsumers == 0 {
+		p.FSubSettings.FQBConsumers = 1
+	}
+
 	if p.FSubSettings.FLogger == nil {
 		p.FSubSettings.FLogger = gopeer_logger.NewLogger(
 			gopeer_logger.NewSettings(&gopeer_logger.SSettings{}),
@@ -102,6 +111,10 @@ func (p *sSettings) GetFetchTimeout() time.Duration {
 
 func (p *sSettings) GetParallel() uint64 {
 	return p.FSubSettings.FParallel
+}
+
+func (p *sSettings) GetQBConsumers() uint64 {
+	return p.FSubSettings.FQBConsumers
 }
 
 func (p *sSettings) GetServiceName() string {